requirements.txt
google-generativeai
openai
anthropic
requests
jsonlines
psutil